LeDao宠物医院管理系统

+++ src/main/java/com/ledao/entity/Goods.java
@@ -73,7 +73,7 @@ public class Goods {
     /**
      * 销售总数
      */
-    private int saleTotal;
+    private Long saleTotal;
 
     @Override
     public String toString() {
